Информационная безопасность баз данных/ЛР4/Выполненная лабораторная/requirements.txt

flask
sqlalchemy
psycopg2-binary
flask-login
python-dotenv
passlib[bcrypt]
